Refactored UriExtensions to be a non-static class.
- Created a QueryString class that wraps up extracting and converting query
  string parameters.
