+++ JabbR/JabbR.csproj
@@ -903,7 +903,7 @@
     <Compile Include="Infrastructure\SafeCollection.cs" />
     <Compile Include="Infrastructure\TaskAsyncHelper.cs" />
     <Compile Include="Infrastructure\TextTransform.cs" />
-    <Compile Include="Infrastructure\UriExtensions.cs" />
+    <Compile Include="Infrastructure\QueryStringCollection.cs" />
     <Compile Include="Migrations\201111290920351_Passwords.cs" />
     <Compile Include="Migrations\201111290920351_Passwords.Designer.cs">
       <DependentUpon>201111290920351_Passwords.cs</DependentUpon>
